[#21][refactor]Board를 board 패키지로 이동
