+++ src/main/java/dev/idion/bladitodo/web/dto/BoardDTO.java
@@ -1,6 +1,6 @@
 package dev.idion.bladitodo.web.dto;
 
-import dev.idion.bladitodo.domain.entity.Board;
+import dev.idion.bladitodo.domain.board.Board;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
