+++ src/main/java/dev/idion/bladitodo/web/dto/BoardWithListIdResponse.java
@@ -1,6 +1,6 @@
 package dev.idion.bladitodo.web.dto;
 
-import dev.idion.bladitodo.domain.entity.Board;
+import dev.idion.bladitodo.domain.board.Board;
 import dev.idion.bladitodo.domain.entity.List;
 import java.util.ArrayList;
 import java.util.stream.Collectors;
